+++ CHANGELOG.md
@@ -7,3 +7,4 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Horizontal scene navigation option that displays scenes as pills across the top of the screen instead of the v13 vertical dropdown
+- Collapsible hotbar that slides down when you want it out of the way